more to go

+++ frontend/android/sm-drinking-system/src/main/java/com/sm/drinking/Drink.java
@@ -4,7 +4,7 @@ package com.sm.drinking;
  * Created by andre on 09.08.13.
  */
 public class Drink {
-    private int id;
+    private long id;
     private String name;
     private int storage_amount;
     private int current;
@@ -18,11 +18,11 @@ public class Drink {
         this.storage_amount = storage_amount;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -60,7 +60,7 @@ public class Drink {
         this.total = total;
     }
 
-    public Drink(int id, int storage_amount, String name, int current, int total, byte[] image) {
+    public Drink(long id, int storage_amount, String name, int current, int total, byte[] image) {
         this.id = id;
         this.storage_amount = storage_amount;
         this.name = name;
